requirements.txt
# Main Dependencies
pandas
numpy
statsmodels
edgartools
click

# tests
pytest
pytz
